+++ src/main/java/com/omgm/admin/kinderGarden/beans/KinderGardenVO.java
@@ -5,19 +5,10 @@ public class KinderGardenVO {
     private String title;
     private String regDate;
 
-    private int no;
-    private String time;
-    private String program;
-
-    public KinderGardenVO() { }
-
-    public KinderGardenVO(int seq, String title, String regDate, int no, String time, String program) {
+    public KinderGardenVO(int seq, String title, String regDate) {
         this.seq = seq;
         this.title = title;
         this.regDate = regDate;
-        this.no = no;
-        this.time = time;
-        this.program = program;
     }
 
     public int getSeq() {
@@ -44,39 +35,12 @@ public class KinderGardenVO {
         this.regDate = regDate;
     }
 
-    public int getNo() {
-        return no;
-    }
-
-    public void setNo(int no) {
-        this.no = no;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
-    }
-
-    public String getProgram() {
-        return program;
-    }
-
-    public void setProgram(String program) {
-        this.program = program;
-    }
-
     @Override
     public String toString() {
         return "KinderGardenVO{" +
                 "seq=" + seq +
                 ", title='" + title + '\'' +
                 ", regDate='" + regDate + '\'' +
-                ", no=" + no +
-                ", time='" + time + '\'' +
-                ", program='" + program + '\'' +
                 '}';
     }
 }
