+++ src/main/java/com/omgm/admin/kinderGarden/service/KinderGardenService.java
@@ -1,10 +1,11 @@
 package com.omgm.admin.kinderGarden.service;
 
+import com.omgm.admin.kinderGarden.beans.KinderGardenRowVO;
 import com.omgm.admin.kinderGarden.beans.KinderGardenVO;
 
 import java.util.List;
 
 public interface KinderGardenService {
     KinderGardenVO getSchedule(KinderGardenVO vo);
-    List<KinderGardenVO> getScheduleRow(KinderGardenVO vo);
+    List<KinderGardenRowVO> getScheduleRow(KinderGardenVO vo);
 }
